[vpr][stats] separate tap util stats
